Add EntityViewerForVariant component
